requirements.txt
Django==4.1.7
ptvsd==4.3.2
drf-nested-routers==0.93.4
djangorestframework==3.14.0
psycopg2-binary==2.9.5
drf-extensions==0.7.1
